tests working, updating wrapasync
