+++ redbox/tests/graph/nodes/test_sends.py
@@ -3,7 +3,6 @@ from unittest.mock import AsyncMock, patch, MagicMock
 from uuid import uuid4
 
 import pytest
-from httpx import ConnectError
 from langchain_core.documents import Document
 from langchain_core.messages import AIMessage, ToolCall
 from langgraph.constants import Send
@@ -17,7 +16,6 @@ from redbox.graph.nodes.sends import (
     no_dependencies,
     run_tools_parallel,
 )
-from redbox.graph.nodes.runner.wrap_async import _get_mcp_headers, wrap_async_tool
 from redbox.graph.nodes.tools import build_search_wikipedia_tool, build_govuk_search_tool
 from redbox.models.chain import DocumentState, RedboxQuery, RedboxState, TaskStatus, configure_agent_task_plan
 from tests.conftest import fake_state
@@ -465,256 +463,3 @@ class TestRunToolsParallelAsync:
 
         response = run_tools_parallel(ai_msg, tools=[tool], state=fake_state)
         assert response is None
-
-
-class TestWrapAsyncTool:
-    def _patch_mcp_env(self, mock_load_tools, mock_http_client, mock_session_class, tools):
-        """Patch MCP networking to allow wrap_async_tool to succeed."""
-        # streamablehttp_client mock
-        mock_read, mock_write = AsyncMock(), AsyncMock()
-        mock_http_cm = AsyncMock()
-        mock_http_cm.__aenter__ = AsyncMock(return_value=(mock_read, mock_write, None))
-        mock_http_cm.__aexit__ = AsyncMock(return_value=None)
-        mock_http_client.return_value = mock_http_cm
-
-        # ClientSession mock
-        mock_session = AsyncMock()
-
-        # initialize() must return something with real strings at serverInfo.name/version
-        mock_server_info = MagicMock()
-        mock_server_info.name = "test-server"
-        mock_server_info.version = "1.0"
-        mock_init_result = MagicMock()
-        mock_init_result.serverInfo = mock_server_info
-        mock_session.initialize = AsyncMock(return_value=mock_init_result)
-
-        mock_session.__aenter__ = AsyncMock(return_value=mock_session)
-        mock_session.__aexit__ = AsyncMock(return_value=None)
-        mock_session_class.return_value = mock_session
-
-        # load_mcp_tools returns the tool
-        mock_load_tools.return_value = tools
-
-        return mock_session
-
-    @pytest.mark.parametrize(
-        "url,expected_exceptions",
-        [
-            ("http://fake-mcp-url", (ConnectError)),  # non-existent hostname
-            ("http://127.0.0.1:59999", (ConnectError)),  # unused localhost port
-        ],
-    )
-    def test_connection_failure(self, fake_mcp_tool, url, expected_exceptions):
-        """Test wrap_async_tool fails when MCP server cannot be reached."""
-        tool = fake_mcp_tool("dummy_tool", return_value=None)
-        tool.metadata["url"] = url
-
-        wrapped = wrap_async_tool(tool, tool.name)
-        args = {"foo": "bar"}
-
-        with pytest.raises(ExceptionGroup) as exc_info:
-            wrapped(args)
-
-        # All inner exceptions should match the expected types
-        exceptions = exc_info.value.exceptions
-        assert all(isinstance(e, expected_exceptions) for e in exceptions)
-
-    @pytest.mark.parametrize("expected_tool_result, expected_documents", MCP_TOOL_RESULTS)
-    @patch("redbox.graph.nodes.runner.wrap_async.ClientSession")
-    @patch("redbox.graph.nodes.runner.wrap_async.streamablehttp_client")
-    @patch("redbox.graph.nodes.runner.wrap_async.load_mcp_tools", new_callable=AsyncMock)
-    def test_returns_expected_results(
-        self,
-        mock_load_tools,
-        mock_http_client,
-        mock_session_class,
-        fake_mcp_tool,
-        expected_tool_result,
-        expected_documents,
-    ):
-        """Test that wrap_async_tool correctly returns results from async tool invocation"""
-        expected_tool_content, expected_tool_metadata = expected_tool_result
-
-        # Mock tool with metadata
-        tool_name = "company_tool"
-        args_schema = {"company_name": {"type": "string"}, "required": ["company_name"]}
-        tool = fake_mcp_tool(tool_name, return_value=expected_tool_content, args_schema=args_schema)
-
-        # mock session with patched mcp setup
-        mock_session = self._patch_mcp_env(mock_load_tools, mock_http_client, mock_session_class, [tool])
-
-        # create the wrapped function
-        wrapped_func = wrap_async_tool(tool, tool_name)
-
-        # rest invocation with sample args
-        test_args = {"company_name": "BMW"}
-        result, metadata = wrapped_func(test_args)
-
-        # verify correct interactions
-        mock_http_client.assert_called_once_with(tool.metadata["url"], headers=None)
-        mock_session.initialize.assert_called_once()
-        mock_load_tools.assert_called_once_with(mock_session)
-        tool.ainvoke.assert_called_once_with(test_args)
-
-        # assert the result matches our expected output
-        assert result == expected_documents
-        assert metadata == expected_tool_metadata
-
-    @pytest.mark.parametrize("expected_tool_result, expected_documents", MCP_TOOL_RESULTS)
-    @patch("redbox.graph.nodes.runner.wrap_async.ClientSession")
-    @patch("redbox.graph.nodes.runner.wrap_async.streamablehttp_client")
-    @patch("redbox.graph.nodes.runner.wrap_async.load_mcp_tools")
-    def test_returns_expected_results_no_args(
-        self,
-        mock_load_tools,
-        mock_http_client,
-        mock_session_class,
-        fake_mcp_tool,
-        expected_tool_result,
-        expected_documents,
-    ):
-        """Test that wrap_async_tool correctly returns results from async tool invocation"""
-        expected_tool_content, expected_tool_metadata = expected_tool_result
-
-        # Mock tool with metadata
-        tool_name = "company_tool"
-        args_schema = {}
-        tool = fake_mcp_tool(tool_name, return_value=expected_tool_content, args_schema=args_schema)
-
-        # mock session with patched mcp setup
-        mock_session = self._patch_mcp_env(mock_load_tools, mock_http_client, mock_session_class, [tool])
-
-        # create the wrapped function
-        wrapped_func = wrap_async_tool(tool, tool_name)
-
-        # rest invocation with sample args
-        test_args = {}
-        result, metadata = wrapped_func(test_args)
-
-        # verify correct interactions
-        mock_http_client.assert_called_once_with(tool.metadata["url"], headers=None)
-        mock_session.initialize.assert_called_once()
-        mock_load_tools.assert_called_once_with(mock_session)
-        tool.ainvoke.assert_called_once_with(test_args)
-
-        # assert the result matches our expected output
-        assert result == expected_documents
-        assert metadata == expected_tool_metadata
-
-    @patch("redbox.graph.nodes.runner.wrap_async.ClientSession")
-    @patch("redbox.graph.nodes.runner.wrap_async.streamablehttp_client")
-    @patch("redbox.graph.nodes.runner.wrap_async.load_mcp_tools", new_callable=AsyncMock)
-    def test_tool_not_found(self, mock_load_tools, mock_http_client, mock_session_class, fake_mcp_tool):
-        """Test wrap_async_tool raises ValueError when the requested tool is not in the MCP tool list."""
-
-        tool = fake_mcp_tool("dummy_tool", return_value=None)
-        wrapped_func = wrap_async_tool(tool, "missing_tool")
-
-        self._patch_mcp_env(mock_load_tools, mock_http_client, mock_session_class, [tool])
-
-        # assert ValueError is raised
-        with pytest.raises(ValueError, match="tool with name 'missing_tool' not found"):
-            wrapped_func({"foo": "bar"})
-
-    def test_sso_token_retrieval_failure(self, fake_mcp_tool):
-        """Test that wrap_async_tool raises when sso_access_token.get() fails."""
-        tool = fake_mcp_tool("dummy_tool", return_value=None)
-        tool.metadata["sso_access_token"] = MagicMock()
-        tool.metadata["sso_access_token"].get.side_effect = RuntimeError("vault unavailable")
-
-        wrapped_func = wrap_async_tool(tool, "dummy_tool")
-
-        with pytest.raises(RuntimeError, match="vault unavailable"):
-            wrapped_func({"foo": "bar"})
-
-    @patch("redbox.graph.nodes.runner.wrap_async.ClientSession")
-    @patch("redbox.graph.nodes.runner.wrap_async.streamablehttp_client")
-    @patch("redbox.graph.nodes.runner.wrap_async.load_mcp_tools")
-    def test_intermediate_step_stripped_when_not_in_schema(
-        self,
-        mock_load_tools,
-        mock_http_client,
-        mock_session_class,
-        fake_mcp_tool,
-    ):
-        """Test that is_intermediate_step is removed from args when not required by the tool schema."""
-        return_value = "some content"
-        tool_name = "company_tool"
-        args_schema = {"company_name": {"type": "string"}, "required": ["company_name"]}
-        tool = fake_mcp_tool(tool_name, return_value=return_value, args_schema=args_schema)
-
-        # make ainvoke return something format_mcp_tool_response can handle, or use a non-datahub type
-        tool.metadata["creator_type"] = MagicMock()  # non-datahub, returns raw result
-        tool.ainvoke = AsyncMock(return_value=return_value)
-
-        self._patch_mcp_env(mock_load_tools, mock_http_client, mock_session_class, [tool])
-
-        wrapped_func = wrap_async_tool(tool, tool_name)
-
-        test_args = {"company_name": "BMW", "is_intermediate_step": True}
-        wrapped_func(test_args)
-
-        # is_intermediate_step should have been popped before ainvoke was called
-        tool.ainvoke.assert_called_once_with({"company_name": "BMW"})
-
-    @patch("redbox.graph.nodes.runner.wrap_async.ClientSession")
-    @patch("redbox.graph.nodes.runner.wrap_async.streamablehttp_client")
-    @patch("redbox.graph.nodes.runner.wrap_async.load_mcp_tools")
-    def test_intermediate_step_retained_when_in_schema(
-        self,
-        mock_load_tools,
-        mock_http_client,
-        mock_session_class,
-        fake_mcp_tool,
-    ):
-        """Test that is_intermediate_step is kept in args when the tool schema requires it."""
-        return_value = "some content"
-        tool_name = "company_tool"
-        args_schema = {"company_name": {"type": "string"}, "required": ["company_name", "is_intermediate_step"]}
-        tool = fake_mcp_tool(tool_name, return_value=return_value, args_schema=args_schema)
-
-        tool.metadata["creator_type"] = MagicMock()
-        tool.ainvoke = AsyncMock(return_value=return_value)
-
-        self._patch_mcp_env(mock_load_tools, mock_http_client, mock_session_class, [tool])
-
-        wrapped_func = wrap_async_tool(tool, tool_name)
-
-        test_args = {"company_name": "BMW", "is_intermediate_step": True}
-        wrapped_func(test_args)
-
-        # is_intermediate_step should be preserved
-        tool.ainvoke.assert_called_once_with({"company_name": "BMW", "is_intermediate_step": True})
-
-    @patch("redbox.graph.nodes.runner.wrap_async.asyncio")
-    def test_asyncio_run_failure(self, mock_asyncio, fake_mcp_tool):
-        """Test that wrap_async_tool re-raises when asyncio.run itself fails."""
-        tool = fake_mcp_tool("dummy_tool", return_value=None)
-        mock_asyncio.run.side_effect = RuntimeError("event loop closed")
-
-        wrapped_func = wrap_async_tool(tool, "dummy_tool")
-
-        with pytest.raises(RuntimeError, match="event loop closed"):
-            wrapped_func({"foo": "bar"})
-
-
-@pytest.mark.parametrize(
-    "token_input, expected_output",
-    [
-        (None, {}),
-        ("", {}),
-        ("   ", {}),
-        ("simple-token-123", {"Authorization": "Bearer simple-token-123"}),
-        ("Bearer already-has-prefix", {"Authorization": "Bearer already-has-prefix"}),
-        ("bearer lowercase-prefix", {"Authorization": "bearer lowercase-prefix"}),
-        ("  token-with-spaces  ", {"Authorization": "Bearer token-with-spaces"}),
-    ],
-)
-def test_get_mcp_headers_logic(token_input, expected_output):
-    """Verify that headers are correctly formatted or returned empty based on input."""
-    assert _get_mcp_headers(token_input) == expected_output
-
-
-def test_get_mcp_headers_no_args():
-    """Verify the default parameter behavior (None)."""
-    assert _get_mcp_headers() == {}
